check for wiki removal

+++ mediawiki/extensions/WikiFarm/src/RemoveWikiJob.php
@@ -25,7 +25,15 @@ class RemoveWikiJob extends \Job {
 
         try {
             echo shell_exec("bash $IP/extensions/WikiFarm/bin/removeWiki.sh wiki$wikiId 2>&1");
-            $this->wikiRepository->removeWiki($wikiId);
+            $resultCheck = shell_exec("bash $IP/extensions/WikiFarm/bin/checkIfWikiExists.sh wiki$wikiId 2> /dev/null");
+            if (strpos($resultCheck, "WIKI") === false
+                && strpos($resultCheck, "DB") === false
+                && strpos($resultCheck, "SOLR") === false
+            ) {
+                $this->wikiRepository->removeWiki($wikiId);
+            } else {
+                wfDebugLog('RemoveWikiJob', "Wiki $wikiId could not be removed completely");
+            }
         } catch(Exception $e) {
             wfDebugLog('RemoveWikiJob', $e->getMessage());
         }
